fix ruby irb
This means we stop linking the rubygems library to the rubygems pkg. Which means our ruby may have an older rubygems gem installed than the rubygems pkg and this is problematic.

at this time I dunno what a solution is since I could not otherwise get irb to work.

Requires a deeper dive but RN I have too many things to fix.